move adel_dynamic to adelwheatdynamic
