+++ test/test_adelwheat_dynamic.py
@@ -1,7 +1,7 @@
 """ Test use cases of adel/fspm wheat coupling"""
 
 
-from alinea.adel.adel_dynamic import AdelWheatDyn
+from alinea.adel.adelwheat_dynamic import AdelWheatDyn
 
 
 def test_add_metamer():
